feat(keymaps): implemented more detailed descriptions (#2282)
* feat(keymaps): implemented more detailed descriptions

By default all keymaps got the description `blink.cmp` for someone new
to the keymaps (like me) it's very helpful to with a tool like
Telescope be able to quickly look up what keymaps do. Hence the
description is very important to provide the detail.

Logic has been added to generate a description in the format:
```
Blink: command_1, command_2, ... (mode)
```

This is just a very quick solution implemented which is good enough for
me, as far as im aware this change is self contained and affects
essentially nothing other than the descriptions.

* style: format lua/blink/cmp/keymap/apply.lua

This is to pass the stylua check

* refactor: simplified description generation

- Instead if using a predefined table, the readable names are instead
generated by simply substituting '_' with space, then captilize the
first letter of each token.

- Changed prefix to instead be `blink.cmp: `

- Removed mode subfix

* spell: Corrected spelling

* refactor: abbreviate custom function description

- Update comment for clarity
- Abbreviate '<Custom Function>' to '<Custom Fn>' to save UI space

* style: minor stylistic changes

---------

Co-authored-by: Liam Dyer <liamcdyer@gmail.com>