fix: documentation draw() function crashing on unmodifiable buffer (#2319)
**Issue:**

The windows that blink.cmp creates were made unmodifiable in
https://github.com/saghen/blink.cmp/pull/2254

This seems to cause issues when using a custom `draw()` function for
rendering custom documentation in a plugin. I worked around it in
https://github.com/mikavilpas/blink-ripgrep.nvim/pull/499 by temporarily
making the buffer modifiable.

But it might still cause issues in other plugins that use custom draw
functions.

**Solution:**

Upstream the fix to blink.cmp: temporarily make the buffer modifiable to
avoid the issue.